[최보근] 채팅 spinner icon 적용, 파일 업로드 UX 고려한 UI 변경
