+++ package.json
@@ -10,7 +10,6 @@
     "@fortawesome/free-regular-svg-icons": "^6.4.2",
     "@fortawesome/free-solid-svg-icons": "^6.4.2",
     "@fortawesome/react-fontawesome": "^0.2.0",
-    "@headlessui/react": "^1.7.18",
     "@mui/icons-material": "^5.14.15",
     "@mui/material": "^5.14.18",
     "@mui/x-data-grid": "^6.18.1",
