+++ src/components/main/Chat/ChatLogs.js
@@ -1,6 +1,6 @@
-import { useSelector } from 'react-redux';
-import TypingAnimation from './TypingAnimation';
-import { Loadicon } from '../../../icons';
+import { useSelector } from "react-redux";
+import TypingAnimation from "./TypingAnimation";
+import { Spinnericon } from "../../../icons";
 
 // 채팅 log 컴포넌트
 function ChatLogs() {
@@ -15,15 +15,19 @@ function ChatLogs() {
         {chatlog.map((message, index) => (
           <li
             key={index}
-            className={`p-3 m-5 rounded-md max-w-2/3 max-h-200 overflow-hidden ${message.user === 'user' ? 'bg-blue-200 ml-auto mr-1' : 'bg-gray-200 ml-1'}`}
+            className={`p-3 m-5 rounded-md max-w-2/3 max-h-200 overflow-hidden ${
+              message.user === "user"
+                ? "bg-blue-200 ml-auto mr-1"
+                : "bg-gray-200 ml-1"
+            }`}
           >
             {<TypingAnimation text={message.message} isNew={message.isNew} />}
           </li>
         ))}
         {loading && (
           <li className="p-3 m-5 rounded-md max-w-2/3 overflow-hidden bg-gray-200 ml-1 flex">
-            <TypingAnimation text={'메시지를 생성 중입니다...'} />
-            <Loadicon />
+            <TypingAnimation text={"메시지를 생성 중입니다..."} />
+            <Spinnericon />
           </li>
         )}
       </ul>
